move shared custom types to a new types.py module
