+++ src/vivarium/framework/lookup/__init__.py
@@ -3,4 +3,5 @@ from vivarium.framework.lookup.manager import (
     LookupTableManager,
     validate_build_table_parameters,
 )
-from vivarium.framework.lookup.table import LookupTable, LookupTableData, ScalarValue
+from vivarium.framework.lookup.table import LookupTable
+from vivarium.types import LookupTableData, ScalarValue
